move collection notification to new mailer; move test
