+++ sorbet/rails-rbi/mailers/collections_mailer.rbi
@@ -1,0 +1,7 @@
+# typed: strong
+# This is an autogenerated file for Rails' mailers.
+# Please rerun bundle exec rake rails_rbi:mailers to regenerate.
+class CollectionsMailer
+  sig { returns(ActionMailer::MessageDelivery) }
+  def self.invitation_to_deposit_email; end
+end
