+++ sorbet/rails-rbi/mailers/notification_mailer.rbi
@@ -9,9 +9,6 @@ class NotificationMailer
   def self.deposited_email; end
 
   sig { returns(ActionMailer::MessageDelivery) }
-  def self.invitation_to_deposit_email; end
-
-  sig { returns(ActionMailer::MessageDelivery) }
   def self.new_version_deposited_email; end
 
   sig { returns(ActionMailer::MessageDelivery) }
